Refactor: unify packet transforms via FramePipeline
Replaces PacketFrameTransforms with direct FramePipeline usage for all packet compression/encryption. Updates FrameReader, FrameSender, and UdpSession to use the new API with explicit parameters. Removes redundant transform helpers from UdpSession. Moves PacketAwaiter to Internal namespace and makes it internal. Adds necessary usings and standardizes exception handling. This streamlines and centralizes packet transformation logic across the SDK.
