+++ src/Nalix.SDK/Transport/Extensions/ControlExtensions.cs
@@ -10,6 +10,7 @@ using Nalix.Common.Networking.Packets;
 using Nalix.Common.Networking.Protocols;
 using Nalix.Framework.DataFrames.SignalFrames;
 using Nalix.Framework.Time;
+using Nalix.SDK.Transport.Internal;
 
 namespace Nalix.SDK.Transport.Extensions;
